Create Q# Oracle 2 balanced case
To construct these two oracles, we need to operate conditionally on the target qubit yyy depending on the value of the data qubit xxx. This means we'll need our conditional gate of choice, CNOT\mathbf{CNOT}CNOT, which is a two-qubit gate. These oracles cannot be implemented using only single-qubit gates.