Create Quantum Parallelism Oracle
To solve the Deutsch problem faster than any classical computer can, we need to prepare the query and target qubits into a superposition state before applying the oracle. This way, the oracle applies linearly to both states in the superposition, and we can take advantage of quantum parallelism:
The action of this phase query for any state xxx input in the first qubit is a conditional phase flip, depending on the value of f(x):f(x):f(x): U∣x,−⟩=(−1)f(x)∣x,−⟩.\mathbf{U} \ket{x, -} = (-1)^{f(x)} \ket{x,-}.U∣x,−⟩=(−1)f(x)∣x,−⟩. Program the gates so that the query qubit is in a uniform superposition and the target qubit is in the ∣−⟩\ket{-}∣−⟩ state. What state should the input qubit ∣x⟩\ket{x}∣x⟩ be in for this query?
 Correct answer: ∣+⟩\ket{+}∣+⟩

In the phase query, the function f(x)f(x)f(x) is evaluated, and its result is applied to a phase flip of the output qubits. To evaluate the function f(x)f(x)f(x) for both possible xxx values in a single query, the query qubit must be prepared in a uniform superposition state, by applying the Hadamard gate before querying the oracle as shown above.