Create DecodeString()
To decode the string, we need to again apply a Hadamard transform to the query qubits. Since the phase query performed a selective phase flip on the terms that matched with the unknown string a\mathbf{a}a, we need to change those phase flips back into bit flips to read out the decoded string.

Then we just need to measure each qubit and output the measurement as a Message, as shown in the operation DecodeString() above.